Google Login Complete

+++ TRAVEL_ON_NA_BACKEND/travel-on-na/src/main/java/com/travelonna/demo/repository/UserTokenRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.travelonna.demo.domain.UserToken;
+import com.travelonna.demo.entity.UserToken;
 
 @Repository
-public interface UserTokenRepository extends JpaRepository<UserToken, Long> {
-    Optional<UserToken> findByUserId(Long userId);
+public interface UserTokenRepository extends JpaRepository<UserToken, Integer> {
+    Optional<UserToken> findByUserId(Integer userId);
     Optional<UserToken> findByRefreshToken(String refreshToken);
 }
